Extract some internal types to native/ - files
