+++ Realm.Shared/Realm.Shared.projitems
@@ -27,8 +27,13 @@
     <Compile Include="$(MSBuildThisFileDirectory)linq\ExpressionVisitor.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)linq\TypeSystem.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MarshalHelpers.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)native\MarshalledVector.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)native\NativeCommon.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)native\NativeException.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)native\NativeTable.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)native\PtrTo.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)native\SchemaObject.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)native\SchemaProperty.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PreserveAttribute.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realm.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)RealmList.cs" />
