+++ Realm.Shared/handles/SchemaHandle.cs
@@ -24,43 +24,11 @@ namespace Realms
 {
     internal class SchemaHandle : RealmHandle
     {
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct Property
-        {
-            [MarshalAs(UnmanagedType.LPStr)]
-            internal string name;
-
-            [MarshalAs(UnmanagedType.U1)]
-            internal Schema.PropertyType type;
-
-            [MarshalAs(UnmanagedType.LPStr)]
-            internal string objectType;
-
-            [MarshalAs(UnmanagedType.I1)]
-            internal bool is_nullable;
-
-            [MarshalAs(UnmanagedType.I1)]
-            internal bool is_primary;
-
-            [MarshalAs(UnmanagedType.I1)]
-            internal bool is_indexed;
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct Object
-        {
-            [MarshalAs(UnmanagedType.LPStr)]
-            internal string name;
-
-            internal int properties_start;
-            internal int properties_end;
-        }
-
         private static class NativeMethods
         {
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "schema_create", CallingConvention = CallingConvention.Cdecl)]
-            public static extern unsafe IntPtr create([MarshalAs(UnmanagedType.LPArray), In] Object[] objects, int objects_length,
-                                                              [MarshalAs(UnmanagedType.LPArray), In] Property[] properties,
+            public static extern unsafe IntPtr create([MarshalAs(UnmanagedType.LPArray), In] SchemaObject[] objects, int objects_length,
+                                                              [MarshalAs(UnmanagedType.LPArray), In] SchemaProperty[] properties,
                                                               [MarshalAs(UnmanagedType.LPArray), Out]IntPtr[] object_schema_handles, out NativeException ex);
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "schema_clone", CallingConvention = CallingConvention.Cdecl)]
@@ -79,7 +47,7 @@ namespace Realms
         {
         }
 
-        public  void Initialize(Object[] objects, int count, Property[] properties, IntPtr[] objectSchemaHandles)
+        public  void Initialize(SchemaObject[] objects, int count, SchemaProperty[] properties, IntPtr[] objectSchemaHandles)
         {
             NativeException nativeException;
             var ptr = NativeMethods.create(objects, count, properties, objectSchemaHandles, out nativeException);
