+++ Realm.Shared/native/SchemaProperty.cs
@@ -1,0 +1,44 @@
+﻿////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2016 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+using System.Runtime.InteropServices;
+
+namespace Realms
+{
+    [StructLayout(LayoutKind.Sequential)]
+    internal struct SchemaProperty
+    {
+        [MarshalAs(UnmanagedType.LPStr)]
+        internal string name;
+
+        [MarshalAs(UnmanagedType.U1)]
+        internal Schema.PropertyType type;
+
+        [MarshalAs(UnmanagedType.LPStr)]
+        internal string objectType;
+
+        [MarshalAs(UnmanagedType.I1)]
+        internal bool is_nullable;
+
+        [MarshalAs(UnmanagedType.I1)]
+        internal bool is_primary;
+
+        [MarshalAs(UnmanagedType.I1)]
+        internal bool is_indexed;
+    }
+}
